requirements.txt
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
plotly==5.22.0
python-dotenv==1.0.1
langchain==0.2.10
langchain-experimental==0.0.63
langchain-openai==0.1.17
openai==1.35.10
